requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Limiter==3.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator==2.1.1
itsdangerous==2.1.2
Pillow==10.2.0